Update Validation annotation documentation
The current docs did not make it clear that annotations that were a part
of the Default group had no affect if the Validation annotation did not
contain the Default group (explicitly or implicitly through
inheritance). See PR #2052 for more information.

This commit also adds test cases to make the current behavior explicit
so any regressions can be caught in the future.