Constraint validation caches error prefixes (#2246)
Previously, once a validation constraint was evaluated the entire string
was cached. The impetus for a cache is that reflection is used to
determine if we need to prefix an error message with additional context
(eg. "query parameter x ..."). Repeatedly calculating this prefix is
wasteful -- thus the cache was born.

The biggest issue with this cache is that while it worked with all
Hibernate Validation default message templates, it failed for message
templates that contained the invalid value (eg. "query parameter x
({validatedValue}) is greater than {value}"), as the same error would be
returned regardless of validatedValue. It would also fail for custom
constraint validators that didn't return the same message everytime.

The fix is to switch the cache from caching the entire error message to
just the error prefix (eg. "query parameter x", "return value", no
prefix).

This has no effect on performance and will allow custom validation
messages in Dropwizard.